omnisight/requirements.txt
openai
pydantic
pydantic-settings
python-dotenv
fastapi
uvicorn
streamlit
pandas
plotly
qdrant-client
langgraph
httpx
orjson
pytest
black
ruff
datasets
pyarrow
huggingface_hub
ucimlrepo
praw
requests
pillow
rapidfuzz
sentence-transformers[image]
pillow